# Core: Fix addon essentials preview preset