--- a/code/builders/builder-vite/src/optimizeDeps.ts
+++ b/code/builders/builder-vite/src/optimizeDeps.ts
@@ -17,7 +17,6 @@ const INCLUDE_CANDIDATES = [
   '@storybook/addon-backgrounds/preview',
   '@storybook/addon-designs/blocks',
   '@storybook/addon-docs/preview',
-  '@storybook/addon-essentials/preview',
   '@storybook/addon-essentials/actions/preview',
   '@storybook/addon-essentials/actions/preview',
   '@storybook/addon-essentials/backgrounds/preview',